Allow mounting with local user
There are permission problems when trying to compile PDFs since the
HOME environment variable is used by texlive to place temporary
files. HOME is now set to the /home/ directory which allows any user
to write.

The example command in the README is updated as well to use a cleaner
docker setup.